\name{RenvCheck-package}
\alias{RenvCheck-package}
\alias{RenvCheck}
\docType{package}
\title{
\packageTitle{RenvCheck}
}
\description{
\packageDescription{RenvCheck}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RenvCheck}
\packageIndices{RenvCheck}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RenvCheck}

Maintainer: \packageMaintainer{RenvCheck}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
